refactor: 重构 decompress 库代码结构
