+++ decompress/tarball_decompressor.py
@@ -20,19 +20,24 @@ class TarballDecompressor(BaseDecompressor):
         :param max_count: 压缩文件解压后，得到文件数量的上限
         :return: None
         """
-        self._verify(path, max_size, max_count)
+        self._verify(path, output_dir, max_size, max_count)
         with tarfile.open(path) as tf:
             tf.extractall(path=output_dir)
 
-    @staticmethod
-    def _verify(path: str, max_size: int, max_count: int) -> NoReturn:
+    def _verify(self, path: str, output_dir: str, max_size: int, max_count: int) -> NoReturn:
         """
-        对压缩文件进行检验，通过设置解压后文件大小上限和文件数量上限，防止压缩炸弹攻击
+        对压缩文件进行检验
         :param path: 压缩文件路径，可以是相对路径或绝对路径
+        :param output_dir: 压缩文件输出目录路径，可以是相对路径或绝对路径
         :param max_size: 压缩文件解压后，得到文件大小的上限
         :param max_count: 压缩文件解压后，得到文件数量的上限
         :return: None
         """
+        # 对压缩文件进行基本的校验
+        self._verify_decompressed_file(path)
+        # 对输出目录进行基本的校验
+        self._verify_output_dir(output_dir)
+        # 校验压缩文件的大小和文件数量是否符合要求
         with tarfile.open(path) as tf:
             total_size = 0
             for _ in range(max_count + 1):
